добавленеи карт клиенту
добавил добавление кредитных и дебетовых карты к аккаунту клиента

+++ bank_forms/src/BankCards/CardManagement.cs
@@ -36,7 +36,7 @@ namespace bank_forms.src.BankCards
         (
             MongoClient client, 
             string validity, 
-            int percent = 0, 
+            double percent = 0, 
             int maxLimit = 0
         )
         {
